+++ src/features/search/tokenizer.ts
@@ -9,6 +9,7 @@ const VALID_FIELDS: Set<FieldSpecifier> = new Set([
   "pmcid",
   "url",
   "keyword",
+  "tag",
 ]);
 
 /**